refactored checkout subtotal component
