+++ frontend/src/components/checkout/checkoutSubtotal/index.js
@@ -1,0 +1,35 @@
+import React from "react";
+import { useHistory } from "react-router-dom";
+import SubTotal from "../checkoutSubtotal";
+import PeopleWhoBought from "../../sitewide/people-who-bought";
+
+function CheckoutSubtotal({ user, basket }) {
+  const history = useHistory();
+
+  if (user?.email && basket.length > 0) {
+    return (
+      <div>
+        <div className="checkout__subtotal">
+          <>
+            <SubTotal />
+            <small className="subtotal__gift">
+              <input type="checkbox" />
+              This order contains a gift
+            </small>
+            <button
+              className="btn-amazon btn-amazon-active f-width"
+              onClick={(e) => history.push("/payment")}
+            >
+              Proceed to Checkout
+            </button>
+          </>
+        </div>
+        <PeopleWhoBought />
+      </div>
+    );
+  } else {
+    return null;
+  }
+}
+
+export default CheckoutSubtotal;
