+++ frontend/src/components/checkout/index.jsx
@@ -1,44 +1,15 @@
 import React from "react";
-import "./checkout.css";
-import { useHistory } from "react-router-dom";
 
 import Cart from "./cart";
-import SubTotal from "./subtotal";
+import CheckoutSubtotal from "./checkoutSubtotal";
+
 import BrowsingHistory from "../sitewide/browsering-history";
-import PeopleWhoBought from "../sitewide/people-who-bought";
 import { useStateValue } from "../../StateProvider";
+
+import "./checkout.css";
 
 function Checkout() {
   const [{ basket, user }, dispatch] = useStateValue();
-
-  const history = useHistory();
-
-  const CheckoutSubtotal = () => {
-    if (user?.email && basket.length > 0) {
-      return (
-        <div>
-          <div className="checkout__subtotal">
-            <>
-              <SubTotal />
-              <small className="subtotal__gift">
-                <input type="checkbox" />
-                This order contains a gift
-              </small>
-              <button
-                className="btn-amazon btn-amazon-active f-width"
-                onClick={(e) => history.push("/payment")}
-              >
-                Proceed to Checkout
-              </button>
-            </>
-          </div>
-          <PeopleWhoBought />
-        </div>
-      );
-    } else {
-      return null;
-    }
-  };
 
   return (
     <>
@@ -51,7 +22,7 @@ function Checkout() {
 
           <Cart />
         </div>
-        <CheckoutSubtotal />
+        <CheckoutSubtotal user={user} basket={basket} />
       </div>
       <BrowsingHistory />
     </>
